refactor: remove dom helper re-export
