--- ui/dom.js
@@ -1,4 +1,0 @@
-export const qs = sel => document.querySelector(sel);
-
-export { setText, setFill, log } from '../src/game/utils.js';
-
